requirements.txt
Flask==0.10.1
Flask-PyMongo==0.4.1
google-api-python-client==1.5.0
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauth2client==2.0.2
paho-mqtt==1.1
pkg-resources==0.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pymongo==3.2.2
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
Werkzeug==0.11.9
